requirements.txt
requests
Pillow
cowsay